requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

